Fix coding like mvc
