+++ controllers/homeController.js
@@ -1,0 +1,8 @@
+const homeController = {
+  home: (req, res) => {
+    const data = {page: 'Home'}
+    res.render('home.ejs', data)
+  }
+}
+
+module.exports = homeController
