+++ controllers/sessionController.js
@@ -1,0 +1,8 @@
+const loginController = {
+  new: (req, res) => {
+    const data = {page: 'Login'}
+    res.render('login.ejs', data)
+  }
+}
+
+module.exports = loginController
